报工提交和报工删除的浮点数修复，修改字段类型

+++ common/dao-api/src/main/java/org/thingsboard/server/dao/mes/dto/OrderProcessRecordSaveDto.java
@@ -3,6 +3,7 @@ package org.thingsboard.server.dao.mes.dto;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -41,7 +42,7 @@ public class OrderProcessRecordSaveDto {
     private String recordUnit;
 
     @ApiModelProperty("盘点数量")
-    private Float recordQty;
+    private BigDecimal recordQty;
 
 //    @ApiModelProperty("盘点数量(手工输入)")
 //    private Float recordManualQty = 0f;
